[FLINK-13041] Move ScheduleMode to flink-core
We need to have this in the core API package so that the Blink Table
Runner can configure the required schedule mode, based on whether we're
executing "streaming" or "batch"-style.
